Downloading the required pdf links
